Results wrapper and BusinessRules added. Command and query refactoring.

+++ Core/ETradeAPI.Application/Features/Products/Commands/DeleteProductByIdCommand.cs
@@ -1,5 +1,9 @@
 ﻿using AutoMapper;
+using ETradeAPI.Application.Features.Products.Rules;
 using ETradeAPI.Application.Repositories.ProductRepository;
+using ETradeAPI.Application.Wrappers.Results.Abstract;
+using ETradeAPI.Application.Wrappers.Results.Concrete;
+using ETradeAPI.Domain.Constants;
 using MediatR;
 using System;
 using System.Collections.Generic;
@@ -9,22 +13,27 @@ using System.Threading.Tasks;
 
 namespace ETradeAPI.Application.Features.Products.Commands
 {
-    public class DeleteProductByIdCommand:IRequest<bool>
+    public class DeleteProductByIdCommand:IRequest<IResult>
     {
         public string Id { get; set; }
-        public class DeleteProductByIdCommandHandler : IRequestHandler<DeleteProductByIdCommand, bool>
+        public class DeleteProductByIdCommandHandler : IRequestHandler<DeleteProductByIdCommand, IResult>
         {
             private readonly IProductWriteRepository _productWriteRepository;
-            private readonly IMapper _mapper;
-            public DeleteProductByIdCommandHandler(IProductWriteRepository productWriteRepository, IMapper mapper)
+            private readonly ProductBusinessRules _productBusinesRules;
+            public DeleteProductByIdCommandHandler(IProductWriteRepository productWriteRepository, ProductBusinessRules productBusinesRules)
             {
                 _productWriteRepository = productWriteRepository;
-                _mapper = mapper;
+                _productBusinesRules = productBusinesRules;
             }
-            public async Task<bool> Handle(DeleteProductByIdCommand request, CancellationToken cancellationToken)
+            public async Task<IResult> Handle(DeleteProductByIdCommand request, CancellationToken cancellationToken)
             {
-                var isDeleted = await _productWriteRepository.RemoveAsync(request.Id);
-                return isDeleted;
+                IResult result = BusinessRules.BusinessRules.Run(await _productBusinesRules.CheckProductIsExist(request.Id));
+                if (result != null)
+                {
+                    return result;
+                }
+                await _productWriteRepository.RemoveAsync(request.Id);
+                return new SuccessResult(Messages.ProductDeleted);
             }
         }
     }
